User Roles

+++ app/controllers/wikis_controller.rb
@@ -9,11 +9,14 @@ class WikisController < ApplicationController
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
+
     @wiki = Wiki.new(params.require(:wiki).permit(:title, :body))
     @wiki.user = current_user
+    authorize @wiki
 
     if @wiki.save
       redirect_to @wiki, notice: "Wiki was added."
@@ -25,10 +28,12 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.update_attributes(params.require(:wiki).permit(:title, :body))
       redirect_to @wiki
@@ -40,6 +45,7 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.destroy
       flash[:notice] = "Wiki was deleted succcessfully."
